You will see

+++ turtlebot.py
@@ -33,6 +33,10 @@ async def on_message(message):
         await client.send_message(message.channel, f"{message.author.mention} This is a christian server!")
         return
 
+    if "scuffed" in message.content.lower():
+        await client.send_message(message.channel, f"{message.author.mention} Scuffed networks is fake and gay!")
+        return
+
     if r"https://discord.gg/" in message.content.lower():
         await client.delete_message(message)
         await client.send_message(message.channel, f"{message.author.mention} No advertizing other discord servers! (If you belive this message is an error, then contact PotatoTurtle#1337)")
